Authorization and policies.

+++ app/Http/Controllers/ThreadController.php
@@ -120,6 +120,8 @@ class ThreadController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('delete', $thread);
+
         $thread->delete();
         if (request()->wantsJson()) {
             return response([], 204);
